Moved parser.py to db_to_graph_parser.py to avoid a Pylance warning on the whole parser.py file
